requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12,<4.0
djangorestframework-simplejwt>=5.0.0,<6.0
dj-rest-auth
django-allauth
requests